% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simKAP2-package.R
\docType{package}
\name{simKAP2-package}
\alias{simKAP2}
\alias{simKAP2-package}
\title{simKAP2: updated simKAP}
\description{
Revised simKAP from the original
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
